Broke down semver module to submodules for readability. See #14
